win32 fixes

+++ socket.hpp
@@ -138,25 +138,25 @@ public:
 
     ev_ssize_t send(const char *buf, std::size_t len, int flags = 0) noexcept
     {
-        return ::send(socket_, buf, len, flags);
+        return ::send(socket_, buf, static_cast<int>(len), flags);
     }
 
     ev_ssize_t recv(char *buf, std::size_t len, int flags = 0) noexcept
     {
-        return ::recv(socket_, buf, len, flags);
+        return ::recv(socket_, buf, static_cast<int>(len), flags);
     }
 
     ev_ssize_t sendto(const ip::addr& addr,
         const char *buf, std::size_t len, int flags = 0) noexcept
     {
-        return ::sendto(socket_, buf, len, flags, addr.sa(), addr.size());
+        return ::sendto(socket_, buf, static_cast<int>(len), flags, addr.sa(), addr.size());
     }
 
     ev_ssize_t recvfrom(sock_addr& sa,
         char *buf, std::size_t len, int flags = 0) noexcept
     {
         ev_socklen_t salen = sock_addr::capacity;
-        auto res = ::recvfrom(socket_, buf, len, flags, sa.sa(), &salen);
+        auto res = ::recvfrom(socket_, buf, static_cast<int>(len), flags, sa.sa(), &salen);
         if (res != code::fail)
             sa.resize(salen);
         return res;
